+++ database/migrations/2019_05_25_161254_create_category_products_details__table.php
@@ -15,8 +15,8 @@ class CreateCategoryProductsDetailsTable extends Migration
     {
         Schema::create('categoryProducts', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('productID');
-            $table->unsignedInteger('categoryID');
+            $table->unsignedBigInteger('productID');
+            $table->unsignedBigInteger('categoryID');
             $table->timestamps();
         });
     }
